+++ src/main/java/com/cgi/silver/bar/marketplace/model/OrderSummary.java
@@ -3,11 +3,12 @@ package com.cgi.silver.bar.marketplace.model;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 @Getter
 @Setter
 public class OrderSummary {
-    private Map<Double, Double> buyOrder;
-    private Map<Double, Double> sellOrder;
+    private Map<BigDecimal, Double> buyOrder;
+    private Map<BigDecimal, Double> sellOrder;
 }
